+++ CHANGELOG.md
@@ -6,8 +6,7 @@ CHANGELOG
 **ENHANCEMENTS**
 - Add a CloudFormation custom resource for creating and managing clusters from CloudFormation.
 - Add `mem_used_percent` and `disk_used_percent` metrics for head node memory and root volume disk utilization tracking on the ParallelCluster CloudWatch dashboard, and set up alarms for monitoring these metrics.
-
-**ENHANCEMENTS**
+- Add ParallelCluster and user's tags in the Launch Templates created by CloudFormation at cluster creation time.
 - Add log rotation support for ParallelCluster managed logs.
 
 **CHANGES**